example_scripts/python/basic/logs_to_trace/requirements.txt
python-dotenv==1.0.1
requests==2.32.4
